Good point to check in.   All of the property tables are hooked up and loading data.   I created the workmanship hackery we discussed.   As always comments and suggestions much appreciated.

+++ Source/ACE.Entity/AceObjectPropertiesBool.cs
@@ -9,7 +9,7 @@ namespace ACE.Entity
         [DbField("AceObjectId", (int)MySqlDbType.UInt32, IsCriteria = true)]
         public uint AceObjectId { get; set; }
 
-        [DbField("boolPropertyId", (int)MySqlDbType.UInt32)]
+        [DbField("boolPropertyId", (int)MySqlDbType.UInt16)]
         public uint BoolPropertyId { get; set; }
 
         [DbField("propertyValue", (int)MySqlDbType.Bit)]
